Master

+++ scripts/install_prerequisites.sh
@@ -25,6 +25,11 @@
 #
 # Note that the root folder MUST be named 'oppia'.
 
+if [ -e "/etc/is_vagrant_vm" ]
+then
+  source $(dirname $0)/vagrant_lockfile.sh || exit 1
+fi
+
 sudo apt-get update
 sudo apt-get install curl
 sudo apt-get install git
@@ -33,4 +38,5 @@ sudo apt-get install python-setuptools
 sudo apt-get install python-dev
 sudo apt-get install python-pip
 sudo apt-get install unzip
+sudo apt-get install python-yaml
 sudo pip install --upgrade pip
